+++ tests/builder/test_json_builder.py
@@ -83,7 +83,7 @@ class TestJsonBuilder (unittest.TestCase):
         self._checkUpType(0, 'OneType', 2, modelTypes, [])
         self._checkUpType(1, 'TwoType', 4, modelTypes, [])
         # TwoType->implicitRef
-        self.assertIsNotNone(modelTypes[1].properties[3].implicitReference)
+        self.assertIsNotNone(modelTypes[1].properties[3].foreignKey)
         self.assertEqual(modelTypes[1].properties[2].type, modelTypes[1].properties[3].implicitReference)
         self._checkUpType(2, 'AnotherType', 2, modelTypes, [])
         self._checkUpType(3, 'DemoEnum', 0, modelTypes, [])